+++ docs/conf.py
@@ -13,12 +13,14 @@
 # serve to show the default.
 
 import sys, os
+import django
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath('..'))
 os.environ['DJANGO_SETTINGS_MODULE'] = 'tests.mockapp.settings'
+django.setup()
 
 # -- General configuration -----------------------------------------------------