+++ tests/test_utils.py
@@ -1,9 +1,8 @@
 from django.core.exceptions import ImproperlyConfigured
+from django.test import override_settings, TestCase
 from django.utils.six.moves import range, reduce
 
 from tidings.utils import collate, import_from_setting
-
-from .base import TestCase, override_settings
 
 
 class MergeTests(TestCase):