+++ tests/mockapp/settings.py
@@ -1,15 +1,4 @@
-import os
-
-import django
-
-
-# Make filepaths relative to settings.
-ROOT = os.path.dirname(os.path.abspath(__file__))
-
-
-def path(*a):
-    return os.path.join(ROOT, *a)
-
+'''Settings for test application.'''
 
 # Django
 DATABASES = {
@@ -27,15 +16,17 @@ INSTALLED_APPS = [
     'tests',
     'tests.mockapp',
 ]
-
-if django.VERSION[:2] < (1, 7):
-    INSTALLED_APPS = INSTALLED_APPS + ['south']
-
 ROOT_URLCONF = 'tests.urls'
+SECRET_KEY = 'yada-yada'
 SITE_ID = 1
-TEST_RUNNER = 'django.test.runner.DiscoverRunner'
-
-MIDDLEWARE_CLASSES = []
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'DIRS': [],
+        'APP_DIRS': True,
+        'OPTIONS': {},
+    },
+]
 
 # Celery
 CELERY_TASK_ALWAYS_EAGER = True
@@ -44,5 +35,3 @@ CELERY_TASK_EAGER_PROPAGATES = True  # Explode loudly during tests.
 # Tidings
 TIDINGS_FROM_ADDRESS = 'nobody@example.com'
 TIDINGS_CONFIRM_ANONYMOUS_WATCHES = True
-
-SECRET_KEY = 'yada-yada'