+++ tidings/migrations/0001_initial.py
@@ -23,8 +23,8 @@ class Migration(migrations.Migration):
                 ('email', models.EmailField(db_index=True, max_length=75, null=True, blank=True)),
                 ('secret', models.CharField(max_length=10, null=True, blank=True)),
                 ('is_active', models.BooleanField(default=False, db_index=True)),
-                ('content_type', models.ForeignKey(blank=True, to='contenttypes.ContentType', null=True)),
-                ('user', models.ForeignKey(blank=True, to=settings.AUTH_USER_MODEL, null=True)),
+                ('content_type', models.ForeignKey(blank=True, to='contenttypes.ContentType', null=True, on_delete=models.CASCADE)),
+                ('user', models.ForeignKey(blank=True, to=settings.AUTH_USER_MODEL, null=True, on_delete=models.CASCADE)),
             ],
             options={
             },
@@ -36,7 +36,7 @@ class Migration(migrations.Migration):
                 ('id', models.AutoField(verbose_name='ID', serialize=False, auto_created=True, primary_key=True)),
                 ('name', models.CharField(max_length=20)),
                 ('value', models.PositiveIntegerField()),
-                ('watch', models.ForeignKey(related_name='filters', to='tidings.Watch')),
+                ('watch', models.ForeignKey(related_name='filters', to='tidings.Watch', on_delete=models.CASCADE)),
             ],
             options={
             },