+++ docs/changes.rst
@@ -3,8 +3,12 @@ Version History
 
 2.0 (To Be Released)
   * Dropped support for Django 1.7 and South.
-  * Migrates Watch.email from a maximum length of 75 to 254, to follow the
+  * Added ``Event.fire(delay=False)``, to avoid using the
+    pickle serializer, which has `security concerns`_.
+  * Migrated Watch.email from a maximum length of 75 to 254, to follow the
     EmailField update in Django 1.8.
+
+.. _`security concerns`: http://docs.celeryproject.org/en/latest/userguide/security.html#serializers
 
 1.2 (2017-03-22)
   * Added support for Django 1.8 and Python 3